+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Crash when not initialized declaration ([#126](https://github.com/esdoc/esdoc/issues/126))
   - Crash when ``@param`` description has ``{@link foo}`` ([#129](https://github.com/esdoc/esdoc/issues/129))
   - Allow particular HTML tags in each descriptions ([#130](https://github.com/esdoc/esdoc/issues/130))
+  - Crash when record + union type is exists ([#152](https://github.com/esdoc/esdoc/issues/152))
 
 ## 0.4.1 (2015-10-18)
 - **Breaking**